requirements.txt
blinker==1.4
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
Deprecated==1.2.13
Flask==2.0.3
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
googleapis-common-protos==1.56.0
greenlet==1.1.2
grpcio==1.44.0
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.0
opentelemetry-api==1.11.1
opentelemetry-exporter-jaeger==1.11.1
opentelemetry-exporter-jaeger-proto-grpc==1.11.1
opentelemetry-exporter-jaeger-thrift==1.11.1
opentelemetry-instrumentation==0.30b1
opentelemetry-instrumentation-flask==0.30b1
opentelemetry-instrumentation-requests==0.30b1
opentelemetry-instrumentation-wsgi==0.30b1
opentelemetry-sdk==1.11.1
opentelemetry-semantic-conventions==0.30b1
opentelemetry-util-http==0.30b1
prometheus-client==0.14.1
protobuf==3.20.1
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.32
SQLAlchemy-Utils==0.38.2
thrift==0.16.0
typing-extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.0.3
wrapt==1.14.0
